+++ lib/src/ui/context_theme_colors.dart
@@ -1,243 +1,5 @@
 import 'package:flutter/material.dart';
 
-@Deprecated(
-    'Use [context.primary] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color primary(context) {
-  return Theme.of(context).colorScheme.primary;
-}
-
-@Deprecated(
-    'Use [context.primaryContainer] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color primaryContainer(context) {
-  return Theme.of(context).colorScheme.primaryContainer;
-}
-
-@Deprecated(
-    'Use [context.onPrimary] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color onPrimary(context) {
-  return Theme.of(context).colorScheme.onPrimary;
-}
-
-@Deprecated(
-    'Use [context.secondary] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color secondary(context) {
-  return Theme.of(context).colorScheme.secondary;
-}
-
-@Deprecated(
-    'Use [context.secondaryContainer] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color secondaryContainer(context) {
-  return Theme.of(context).colorScheme.secondaryContainer;
-}
-
-@Deprecated(
-    'Use [context.onSecondary] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color onSecondary(context) {
-  return Theme.of(context).colorScheme.onSecondary;
-}
-
-@Deprecated(
-    'Use [context.background] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color background(context) {
-  return Theme.of(context).colorScheme.background;
-}
-
-@Deprecated(
-    'Use [context.onBackground] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color onBackground(context) {
-  return Theme.of(context).colorScheme.onBackground;
-}
-
-@Deprecated(
-    'Use [context.surface] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color surface(context) {
-  return Theme.of(context).colorScheme.surface;
-}
-
-@Deprecated(
-    'Use [context.onSurface] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color onSurface(context) {
-  return Theme.of(context).colorScheme.onSurface;
-}
-
-@Deprecated(
-    'Use [context.error] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color error(context) {
-  return Theme.of(context).colorScheme.error;
-}
-
-@Deprecated(
-    'Use [context.onError] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color onError(context) {
-  return Theme.of(context).colorScheme.onError;
-}
-
-@Deprecated(
-    'Use [context.backgroundColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color backgroundColor(context) {
-  return Theme.of(context).backgroundColor;
-}
-
-@Deprecated(
-    'Use [context.bottomAppBarColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color bottomAppBarColor(context) {
-  return Theme.of(context).bottomAppBarColor;
-}
-
-@Deprecated(
-    'Use [context.canvasColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color canvasColor(context) {
-  return Theme.of(context).canvasColor;
-}
-
-@Deprecated(
-    'Use [context.cardColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color cardColor(context) {
-  return Theme.of(context).cardColor;
-}
-
-@Deprecated(
-    'Use [context.dialogBackgroundColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color dialogBackgroundColor(context) {
-  return Theme.of(context).dialogBackgroundColor;
-}
-
-@Deprecated(
-    'Use [context.disabledColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color disabledColor(context) {
-  return Theme.of(context).disabledColor;
-}
-
-@Deprecated(
-    'Use [context.dividerColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color dividerColor(context) {
-  return Theme.of(context).dividerColor;
-}
-
-@Deprecated(
-    'Use [context.errorColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color errorColor(context) {
-  return Theme.of(context).errorColor;
-}
-
-@Deprecated(
-    'Use [context.focusColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color focusColor(context) {
-  return Theme.of(context).focusColor;
-}
-
-@Deprecated(
-    'Use [context.highlightColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color highlightColor(context) {
-  return Theme.of(context).highlightColor;
-}
-
-@Deprecated(
-    'Use [context.hintColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color hintColor(context) {
-  return Theme.of(context).hintColor;
-}
-
-@Deprecated(
-    'Use [context.hoverColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color hoverColor(context) {
-  return Theme.of(context).hoverColor;
-}
-
-@Deprecated(
-    'Use [context.indicatorColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color indicatorColor(context) {
-  return Theme.of(context).indicatorColor;
-}
-
-@Deprecated(
-    'Use [context.scaffoldBackgroundColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color scaffoldBackgroundColor(context) {
-  return Theme.of(context).scaffoldBackgroundColor;
-}
-
-@Deprecated(
-    'Use [context.secondaryHeaderColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color secondaryHeaderColor(context) {
-  return Theme.of(context).secondaryHeaderColor;
-}
-
-@Deprecated(
-    'Use [context.shadowColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color shadowColor(context) {
-  return Theme.of(context).shadowColor;
-}
-
-@Deprecated(
-    'Use [context.selectedRowColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color selectedRowColor(context) {
-  return Theme.of(context).selectedRowColor;
-}
-
-@Deprecated(
-    'Use [context.splashColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color splashColor(context) {
-  return Theme.of(context).splashColor;
-}
-
-@Deprecated(
-    'Use [context.toggleableActiveColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color toggleableActiveColor(context) {
-  return Theme.of(context).toggleableActiveColor;
-}
-
-@Deprecated(
-    'Use [context.unselectedWidgetColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color unselectedWidgetColor(context) {
-  return Theme.of(context).unselectedWidgetColor;
-}
-
-@Deprecated(
-    'Use [context.surfaceTint] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color surfaceTintColor(context) {
-  return Theme.of(context).colorScheme.surfaceTint;
-}
-
-@Deprecated(
-    'Use [context.backgroundColor] instead. This will be removed in the next major version.'
-    'This feature was deprecated after v0.0.13+2')
-Color? appBarColor(context) {
-  return Theme.of(context).appBarTheme.backgroundColor;
-}
-
 extension ContextThemeColors on BuildContext {
   Color get primary => Theme.of(this).colorScheme.primary;
   Color get primaryContainer => Theme.of(this).colorScheme.primaryContainer;
@@ -251,14 +13,14 @@ extension ContextThemeColors on BuildContext {
   Color get onSurface => Theme.of(this).colorScheme.onSurface;
   Color get error => Theme.of(this).colorScheme.error;
   Color get onError => Theme.of(this).colorScheme.onError;
-  Color get backgroundColor => Theme.of(this).backgroundColor;
-  Color get bottomAppBarColor => Theme.of(this).bottomAppBarColor;
+  Color get bottomAppBarColor =>
+      BottomAppBarTheme.of(this).color ?? Theme.of(this).bottomAppBarColor;
   Color get canvasColor => Theme.of(this).canvasColor;
   Color get cardColor => Theme.of(this).cardColor;
   Color get dialogBackgroundColor => Theme.of(this).dialogBackgroundColor;
   Color get disabledColor => Theme.of(this).disabledColor;
   Color get dividerColor => Theme.of(this).dividerColor;
-  Color get errorColor => Theme.of(this).errorColor;
+  Color get errorColor => Theme.of(this).colorScheme.error;
   Color get focusColor => Theme.of(this).focusColor;
   Color get highlightColor => Theme.of(this).highlightColor;
   Color get hintColor => Theme.of(this).hintColor;
@@ -267,8 +29,14 @@ extension ContextThemeColors on BuildContext {
   Color get scaffoldBackgroundColor => Theme.of(this).scaffoldBackgroundColor;
   Color get secondaryHeaderColor => Theme.of(this).secondaryHeaderColor;
   Color get shadowColor => Theme.of(this).shadowColor;
+  @Deprecated(
+      'Use [context.selectedRowColor] instead. This will be removed in the next major version.'
+      'This feature was deprecated after v0.0.20')
   Color get selectedRowColor => Theme.of(this).selectedRowColor;
   Color get splashColor => Theme.of(this).splashColor;
+  @Deprecated(
+      'Use [context.toggleableActiveColor] instead. This will be removed in the next major version.'
+      'This feature was deprecated after v0.0.20')
   Color get toggleableActiveColor => Theme.of(this).toggleableActiveColor;
   Color get unselectedWidgetColor => Theme.of(this).unselectedWidgetColor;
   Color get surfaceTintColor => Theme.of(this).colorScheme.surfaceTint;